requirements.txt
mcp
psycopg2-binary
pgvector
pydantic
numpy
sentence-transformers
